+++ index.js
@@ -1,0 +1,11 @@
+const contactForm = document.getElementById('contact-form');
+
+if (contactForm) {
+    contactForm.addEventListener(
+        'submit',
+        function submit (event) {
+            event.preventDefault();
+            this.reset();
+        }
+    );
+}